+++ CHANGELOG.md
@@ -35,6 +35,7 @@ Please mark all change in change log and use the issue from GitHub
 - \#908 - Move "primary_path" and "secondary_path" to storage config
 - \#931 - Remove "collector" from config
 - \#966 - Update NOTICE.md
+- \#1078 - Move 'insert_buffer_size' to Cache Config section
 
 ## Task